+++ src/getexchange.c
@@ -31,7 +31,7 @@
 #include <ctype.h>
 
 #include "addspot.h"
-#include "callinput.h"
+#include "audio.h"
 #include "cw_utils.h"
 #include "change_rst.h"
 #include "globalvars.h"
