+++ src/globalvars.h
@@ -253,7 +253,6 @@ extern bool debugflag;
 extern bool trx_control;
 extern bool nopacket;
 extern bool verbose;
-extern bool vk_running;
 
 extern double DEST_Lat;
 extern double DEST_Long;
